api/requirements.txt
pandas
numpy
scikit-learn
flask
flask_cors
uvicorn
gunicorn
pyspark
pymongo[srv]
google-auth-oauthlib
google-api-python-client